Move code_id_for_object
